extract task flow from item-by-id component to simplify it
